refactor: Move package build + publish logic to registry
Moves the logic for converting a wasmer.toml package file into a tar.gz
and publishing it to the backend into the wasmer_registry crate.
