+++ lib/registry/src/package/sql/queries/insert_interface.sql
