+++ lib/registry/src/package/sql/queries/wasm_interface_existence_check.sql
